hiding delete group btn if user is not admin

+++ project/app/src/main/java/com/example/sport_app_client/groupActivities/FBGroupFragment.java
@@ -192,7 +192,11 @@ public class FBGroupFragment extends Fragment implements GameCreatedListener, Ga
         leaveGroupBTN.setOnClickListener(view -> onLeaveGroupBTNClick());
 
         this.deleteGroupBTN = view.findViewById(R.id.groupSettingsDeleteBTN);
-        deleteGroupBTN.setOnClickListener(view -> onDeleteGroupBTNClick());
+        if (MyGlobals.associatedFBMember.getIsAdmin()) {
+            deleteGroupBTN.setOnClickListener(view -> onDeleteGroupBTNClick());
+        } else {
+            deleteGroupBTN.setVisibility(View.GONE);
+        }
     }
 
     /**
